+++ player/src/main/java/com/tpstream/player/TpStreamPlayer.kt
@@ -6,7 +6,7 @@ import android.os.Looper
 import com.google.common.collect.ImmutableList
 import com.tpstream.player.data.Video
 import com.tpstream.player.data.VideoRepository
-import com.tpstream.player.data.source.network.NetworkClient
+import com.tpstream.player.util.NetworkClient
 import com.tpstream.player.offline.DownloadTask
 import com.tpstream.player.offline.VideoDownload
 import com.tpstream.player.offline.VideoDownloadManager
