+++ player/src/test/java/com/tpstream/player/NetworkTest.kt
@@ -2,7 +2,7 @@ package com.tpstream.player
 
 import androidx.arch.core.executor.testing.InstantTaskExecutorRule
 import com.tpstream.player.data.source.network.NetworkAsset
-import com.tpstream.player.data.source.network.NetworkClient
+import com.tpstream.player.util.NetworkClient
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.runBlocking
 import okhttp3.mockwebserver.MockResponse
